requirements.txt
absl-py==1.4.0
accelerate==0.27.2
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
# awscli==1.29.46
backoff==1.11.1
bio==1.6.0
biopython==1.82
biothings-client==0.3.1
bleach==6.1.0
# botocore==1.34.113
botocore==1.31.46
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.4.1
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.8.0
conda-pack==0.7.1
contourpy==1.1.1
cycler==0.12.1
datasets==2.14.5
dill==0.3.7
docker-pycreds==0.4.0
docutils==0.16
einops==0.7.0
environs==11.0.0
et-xmlfile==1.1.0
evaluate==0.4.0
filelock==3.12.4
fonttools==4.47.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.0
google-auth-oauthlib==1.0.0
gprofiler-official==1.0.0
grad-cam==1.5.0
graphviz==0.20.1
greenlet==3.0.3
grpcio==1.58.0
huggingface-hub==0.21.1
humanfriendly==10.0
humanize==4.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
kaggle==1.6.14
kiwisolver==1.4.5
lit==16.0.6
Mako==1.3.0
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.21.2
matplotlib==3.7.4
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
multiprocessing-logging==0.3.4
mygene==3.2.2
networkx==3.1
ninja==1.11.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.76
openpyxl==3.1.2
optuna==3.5.0
packaging==24.0
pandas==2.0.3
patsy==0.5.5
peft==0.5.0
# petrel-oss-sdk===v2.2.1-2-g1505ef3-master
Pillow==10.1.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.24.3
psutil==5.9.5
ptvsd==4.3.2
pyahocorasick==2.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyfaidx==0.8.0
pyparsing==3.1.1
pypng==0.20220715.0
pytabix==0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.7.2
# s3transfer==0.10.1
s3transfer==0.6
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.1
sentry-sdk==1.43.0
setproctitle==1.3.3
sigopt==8.8.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
statsmodels==0.14.1
stripedhyena==0.2.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.2.0
tokenizers==0.15.2
# torch==1.13.1+cu117
# torch-summary==1.4.5
# torchaudio==0.13.1+cu117
# torchvision==0.14.1+cu117
# torchviz==0.0.2
# tqdm==4.66.1
transformers==4.38.1
# triton==2.0.0.dev20221202
triton==2.0.0
ttach==0.0.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.18
wandb==0.16.4
webencodings==0.5.1
Werkzeug==2.3.7
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2
